Refactor Entries.tsx
