+++ src/views/entries/styles.ts
@@ -1,0 +1,35 @@
+export const styles = {
+  outerDiv: {
+    height: 'inherit',
+    overflowY: 'auto',
+    width: '100%',
+  },
+
+  ul: {
+    margin: '0px',
+    listStyleType: 'none',
+    padding: '0px',
+  },
+
+  fixedPosition: {
+    position: 'fixed',
+    width: '342px',
+    top: '0px',
+  },
+
+  listHeader: {
+    display: 'flex',
+    width: '347px',
+    height: '27px',
+    background: '#F4F4F4',
+    color: '#B9B0B0',
+    paddingLeft: '5px',
+    alignItems: 'center',
+  },
+
+  listItems: {
+    color: '#908F8F',
+    paddingleft: '10px',
+    marginLeft: '10px',
+  },
+};
